requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
decorator==4.1.2
Django==1.11.4
django-extensions==1.9.1
django-templated-mail==1.0.0
djangorestframework==3.7.3
enum34==1.1.6
httpie==0.9.9
idna==2.6
ipython==5.4.1
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
